+++ HISTORY.md
@@ -1,12 +1,14 @@
 # History
 
-## 0.10.0 (Unreleased)
+## 0.10.0 (2023-10-07)
+* Add support for conditional requests with `ETag` and `Last-Modified`
+* If a DynamoDB item exceeds the max size (400KB), skip writing to the cache and log a warning instead of raising an error
 * Add `CachedResponse.closed` attribute for compatibility with `aiohttp.ClientResponse`
 * Close `aiosqlite` thread if it's still running when session object is deleted
 * Move redirects cache for `FileBackend` into same directory as cached response files
 * Fix issue in which `CachedSession.disabled()` prevents only cache read but not write
 
-## 0.9.1 (2023-09-20)
+### 0.9.1 (2023-09-20)
 * Remove unintended optional dependencies in both PyPI and conda-forge packages
 
 ## 0.9.0 (2023-09-19)